handle json docs containing an object per line
Large datasets in JSON sometimes come in files with
an object entry per line. When looking for the schema, the tool
could quickly list enough entries to fill the context window.

This updates the tool to try and process larger files with
the slurp functionality and treat all entries in the file as items
in an array. This lets lenght functions work, and grabbing a single
entry.

Signed-off-by: Bill Maxwell <cloudnautique@users.noreply.github.com>